requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.30,<3.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
jinja2>=3.1.4,<4.0.0
apscheduler>=3.10.4,<4.0.0
httpx>=0.27.0,<1.0.0
openai>=1.40.0,<2.0.0
Pillow>=10.4.0,<11.0.0
python-multipart>=0.0.9,<1.0.0
google-api-python-client>=2.141.0,<3.0.0
google-auth>=2.34.0,<3.0.0
google-auth-oauthlib>=1.2.1,<2.0.0
PyYAML>=6.0.2,<7.0.0
pytest>=8.3.0,<9.0.0

